requirements.txt
# Flask==3.1.0
# requests==2.32.3
# bs4==0.0.2
# langchain==0.3.7
# python-dotenv==1.0.1
# langchain-community==0.3.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.56.1
orjson==3.10.12
packaging==24.2
pip==24.2
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
Flask-Cors==5.0.0